+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## 2016-01-12
 
 ### Added
+- Added arrToCSV() function.
+
+## 2016-01-12
+
+### Added
 - build.sh: fixed typo that was preventing the build number from incrementing.
 - Added main source file and converted the functions to TypeScript.
 - Initial commit.